requirements.txt
flask
flask_apscheduler
flask_login
flask_migrate
flask-restful
flask_sqlalchemy
flask_wtf
napalm
netmiko
nornir
jinja2
simplekml
tacacs_plus
wtforms
xlrd
pyyaml
passlib
gunicorn